restructure evaluator code
